app/player.py
import random

random.seed(123)

pearson_table = list(range(256))
random.shuffle(pearson_table)


class Player:
    def __init__(self, uid: str, name: str, score: int=0) -> None:
        self._uid = uid
        self._name = name
        self._score = score

    @property
    def name(self):
        return self._name

    @property
    def uid(self) -> str:
        return self._uid

    @property
    def score(self) -> int:
        return self._score

    @score.setter
    def score(self, score: int) -> None:
        self._score = score

    def __eq__(self, other) -> bool:
        return self.score == other.score

    def __ge__(self, other) -> bool:
        return self.score >= other.score

    def __lt__(self, other) -> bool:
        return self.score < other.score

    @staticmethod
    def quicksort(arr: list) -> list:
        if len(arr) < 2:
            return arr
        pivot = arr[0]
        less_than_pivot = [v for v in arr[1:] if v < pivot]
        greater_than_pivot = [v for v in arr[1:] if v >= pivot]
        return Player.quicksort(greater_than_pivot) + [pivot] + Player.quicksort(less_than_pivot)


    def __str__(self):
        return f"{self.uid}, {self.name}, {self.score}"

    def __repr__(self):
        return f"{self.uid}, {self.name}, {self.score}"

    @staticmethod
    def hash(key: str, size: int) -> int:
        hash_ = 0
        for char in key:
            hash_ = pearson_table[hash_ ^ ord(char)]
        return hash_ % size

    # https://stackoverflow.com/questions/2909106/whats-a-correct-and-good-way-to-implement-hash
    def __key(self):
        return self._uid, self._name

    def __hash__(self):
        return self.hash(self.uid, len(Player.__key(self)))

    # def __eq__(self, other):
    #     if isinstance(other, Player):
    #         return self.uid == other.uid
    #     return NotImplemented

